Decoupling using events

+++ Assets/scripts/Bullets/Model/ModelBullet.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class ModelBullet 
 {
-    public float bodyDamage = 10;
-    public float criticalDamage = 30;
+    public int bodyDamage = 10;
+    public int criticalDamage = 30;
     public float power = 10;
     public float lifeTime = 3;
     public bool explosion = false;
